refactor(api): move server bootstrapping and error handling to presentation layer
- Created dedicated express folder in presentation layer
- Moved app and server initialization logic from index.ts to server.ts
- Moved error middleware to presentation/express/middlewares
- Updated imports and references to logger
