+++ apps/express-app/src/presentation/express/middlewares/index.ts
@@ -1,0 +1,1 @@
+export * from "./errors";
